requirements.txt
setuptools~=40.8.0
psycopg2-binary~=2.8.5
postgis~=1.0.4
